+++ airbyte-integrations/connectors/source-s3/main.py
@@ -3,40 +3,7 @@
 #
 
 
-import sys
-import traceback
-from datetime import datetime
-from typing import List
-
-from airbyte_cdk.entrypoint import AirbyteEntrypoint, launch
-from airbyte_cdk.models import AirbyteErrorTraceMessage, AirbyteMessage, AirbyteTraceMessage, TraceType, Type
-from source_s3.v4 import Config, Cursor, SourceS3, SourceS3StreamReader
-
-
-def get_source(args: List[str]):
-    catalog_path = AirbyteEntrypoint.extract_catalog(args)
-    try:
-        return SourceS3(SourceS3StreamReader(), Config, catalog_path, cursor_cls=Cursor)
-    except Exception:
-        print(
-            AirbyteMessage(
-                type=Type.TRACE,
-                trace=AirbyteTraceMessage(
-                    type=TraceType.ERROR,
-                    emitted_at=int(datetime.now().timestamp() * 1000),
-                    error=AirbyteErrorTraceMessage(
-                        message="Error starting the sync. This could be due to an invalid configuration or catalog. Please contact Support for assistance.",
-                        stack_trace=traceback.format_exc(),
-                    ),
-                ),
-            ).json()
-        )
-        return None
-
+from source_s3.run import run
 
 if __name__ == "__main__":
-    _args = sys.argv[1:]
-    source = get_source(_args)
-
-    if source:
-        launch(source, _args)
+    run()
